+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 5.10-SNAPSHOT
 -------------
 
+ADDED:
+
+- new decorator class for JMSpinner that encapsulate a property model and bind it to it
+
 Version 5.9
 -------------